CheckPoint system finihsed
New checkpoint sysem has been created. Refacrorization in GameManager and PlayerMovement script.
