+++ Cube Dash/Assets/__Scripts/Player/SpawnManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e68336fe016365b4385888356118c7af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
