+++ Cube Dash/UserSettings/EditorUserSettings.asset
@@ -15,22 +15,22 @@ EditorUserSettings:
       value: 5a05075007060c02590a0f21402607444116497f752a7f6374284a31e3b8323e
       flags: 0
     RecentlyUsedSceneGuid-3:
-      value: 00090156005051585e0c5d7613200744434e1d7d29797e66757c4a30e4b5626f
-      flags: 0
-    RecentlyUsedSceneGuid-4:
       value: 5755515752070f5a5e0a5d7711270744414f4c792f2976677c284c30bbe2623c
       flags: 0
-    RecentlyUsedSceneGuid-5:
+    RecentlyUsedSceneGuid-4:
       value: 05090c5f57010a0c095d0a74487b0a44444e1e2e287970332b7c4d60e3e4673c
       flags: 0
-    RecentlyUsedSceneGuid-6:
+    RecentlyUsedSceneGuid-5:
       value: 5602525e06530a5e0e5f5a73487a0b4441151a2b7f2d7533792a4961b3b33769
       flags: 0
-    RecentlyUsedSceneGuid-7:
+    RecentlyUsedSceneGuid-6:
       value: 07010c5f55575e5d5a5c5a7016755b44104f402b782a203328281b32b6b2313d
       flags: 0
-    RecentlyUsedSceneGuid-8:
+    RecentlyUsedSceneGuid-7:
       value: 0604525e000150025b085d76417a5944444f4a7b7578226028784464b1b36c68
+      flags: 0
+    RecentlyUsedSceneGuid-8:
+      value: 00090156005051585e0c5d7613200744434e1d7d29797e66757c4a30e4b5626f
       flags: 0
     RecentlyUsedSceneGuid-9:
       value: 51000507540d5e5d55585a7249735b44404e4a7f292e25362c7d196bb1e33761
